--- src/pages/Routes.ts
@@ -1,15 +1,0 @@
-export const APP_ROUTE = '/app';
-
-export const ROUTES = {
-
-  // Rutas de PublicLayout  
-  home: '/',
-  
-  // Rutas de AppLayout
-  app: APP_ROUTE,
-  editor: `${APP_ROUTE}/editor`,
-  login: `${APP_ROUTE}/login`,
-  signin: `${APP_ROUTE}/signin`,
-  profile: `${APP_ROUTE}/profile`,
-
-};
